common.subprocessing: split subprocessing into separate files
